Finalizado v1 para o issue Adicionar Eventos #13
