+++ src/main/resources/application.properties
@@ -7,3 +7,5 @@ quarkus.datasource.jdbc.min-size=2
 
 quarkus.hibernate-orm.database.generation=drop-and-create
 quarkus.hibernate-orm.log.sql=false
+
+quarkus.http.cors=true
